Two different text displays created
script.rpy shows an example of how it works. I'm using a global variable called "say_style" that alternates between Interrogation and Cutscene mode.  say_style defaults in screens.rpy. I have yet to adjust the max line width size for the different textboxes, so I'll try to do that tonight. Currently things are hardcoded for 1920x1080, which I recognize is not ideal, but this was more of a test to familiarize myself with the software. Apologies if this causes problems during development, but after a quick glimpse of how renpy handles screen resolution I think we should be fine. Just thought that it would be worth mentioning.